Move 'DEFAULT_RETRY' (w/ its predicate) to a new public 'retry' module.
Publish it, w/ its docstring, in the reference docs.
