+++ bigquery/google/cloud/bigquery/job.py
@@ -22,7 +22,6 @@ from six.moves import http_client
 import google.api_core.future.polling
 from google.cloud import exceptions
 from google.cloud.exceptions import NotFound
-from google.cloud._helpers import _datetime_from_microseconds
 from google.cloud.bigquery.dataset import DatasetReference
 from google.cloud.bigquery.external_config import ExternalConfig
 from google.cloud.bigquery.query import _query_param_from_api_repr
@@ -30,14 +29,13 @@ from google.cloud.bigquery.query import ArrayQueryParameter
 from google.cloud.bigquery.query import ScalarQueryParameter
 from google.cloud.bigquery.query import StructQueryParameter
 from google.cloud.bigquery.query import UDFResource
+from google.cloud.bigquery.retry import DEFAULT_RETRY
 from google.cloud.bigquery.schema import SchemaField
 from google.cloud.bigquery.table import EncryptionConfiguration
 from google.cloud.bigquery.table import TableReference
 from google.cloud.bigquery.table import Table
 from google.cloud.bigquery.table import TimePartitioning
 from google.cloud.bigquery import _helpers
-from google.cloud.bigquery._helpers import DEFAULT_RETRY
-from google.cloud.bigquery._helpers import _int_or_none
 
 _DONE_STATE = 'DONE'
 _STOPPED_REASON = 'stopped'
@@ -379,7 +377,7 @@ class _AsyncJob(google.api_core.future.polling.PollingFuture):
         if statistics is not None:
             millis = statistics.get('creationTime')
             if millis is not None:
-                return _datetime_from_microseconds(millis * 1000.0)
+                return _helpers._datetime_from_microseconds(millis * 1000.0)
 
     @property
     def started(self):
@@ -392,7 +390,7 @@ class _AsyncJob(google.api_core.future.polling.PollingFuture):
         if statistics is not None:
             millis = statistics.get('startTime')
             if millis is not None:
-                return _datetime_from_microseconds(millis * 1000.0)
+                return _helpers._datetime_from_microseconds(millis * 1000.0)
 
     @property
     def ended(self):
@@ -405,7 +403,7 @@ class _AsyncJob(google.api_core.future.polling.PollingFuture):
         if statistics is not None:
             millis = statistics.get('endTime')
             if millis is not None:
-                return _datetime_from_microseconds(millis * 1000.0)
+                return _helpers._datetime_from_microseconds(millis * 1000.0)
 
     def _job_statistics(self):
         """Helper for job-type specific statistics-based properties."""
@@ -923,7 +921,7 @@ class LoadJobConfig(_JobConfig):
         See
         https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs#configuration.load.skipLeadingRows
         """
-        return _int_or_none(self._get_sub_prop('skipLeadingRows'))
+        return _helpers._int_or_none(self._get_sub_prop('skipLeadingRows'))
 
     @skip_leading_rows.setter
     def skip_leading_rows(self, value):
@@ -1876,7 +1874,7 @@ class QueryJobConfig(_JobConfig):
         See
         https://g.co/cloud/bigquery/docs/reference/rest/v2/jobs#configuration.query.maximumBytesBilled
         """
-        return _int_or_none(self._get_sub_prop('maximumBytesBilled'))
+        return _helpers._int_or_none(self._get_sub_prop('maximumBytesBilled'))
 
     @maximum_bytes_billed.setter
     def maximum_bytes_billed(self, value):
@@ -2361,7 +2359,7 @@ class QueryJob(_AsyncJob):
     @property
     def slot_millis(self):
         """Union[int, None]: Slot-milliseconds used by this query job."""
-        return _int_or_none(self._job_statistics().get('totalSlotMs'))
+        return _helpers._int_or_none(self._job_statistics().get('totalSlotMs'))
 
     @property
     def statement_type(self):
@@ -2606,7 +2604,7 @@ class QueryPlanEntry(object):
         """Union[Datetime, None]: Datetime when the stage started."""
         if self._properties.get('startMs') is None:
             return None
-        return _datetime_from_microseconds(
+        return _helpers._datetime_from_microseconds(
                 int(self._properties.get('startMs')) * 1000.0)
 
     @property
@@ -2614,7 +2612,7 @@ class QueryPlanEntry(object):
         """Union[Datetime, None]: Datetime when the stage ended."""
         if self._properties.get('endMs') is None:
             return None
-        return _datetime_from_microseconds(
+        return _helpers._datetime_from_microseconds(
                 int(self._properties.get('endMs')) * 1000.0)
 
     @property
@@ -2622,7 +2620,7 @@ class QueryPlanEntry(object):
         """List(int): Entry IDs for stages that were inputs for this stage."""
         if self._properties.get('inputStages') is None:
             return []
-        return [_int_or_none(entry)
+        return [_helpers._int_or_none(entry)
                 for entry in self._properties.get('inputStages')]
 
     @property
@@ -2630,26 +2628,27 @@ class QueryPlanEntry(object):
         """Union[int, None]: Number of parallel input segments within
         the stage.
         """
-        return _int_or_none(self._properties.get('parallelInputs'))
+        return _helpers._int_or_none(self._properties.get('parallelInputs'))
 
     @property
     def completed_parallel_inputs(self):
         """Union[int, None]: Number of parallel input segments completed."""
-        return _int_or_none(self._properties.get('completedParallelInputs'))
+        return _helpers._int_or_none(
+            self._properties.get('completedParallelInputs'))
 
     @property
     def wait_ms_avg(self):
         """Union[int, None]: Milliseconds the average worker spent waiting to
         be scheduled.
         """
-        return _int_or_none(self._properties.get('waitMsAvg'))
+        return _helpers._int_or_none(self._properties.get('waitMsAvg'))
 
     @property
     def wait_ms_max(self):
         """Union[int, None]: Milliseconds the slowest worker spent waiting to
         be scheduled.
         """
-        return _int_or_none(self._properties.get('waitMsMax'))
+        return _helpers._int_or_none(self._properties.get('waitMsMax'))
 
     @property
     def wait_ratio_avg(self):
@@ -2672,14 +2671,14 @@ class QueryPlanEntry(object):
         """Union[int, None]: Milliseconds the average worker spent reading
         input.
         """
-        return _int_or_none(self._properties.get('readMsAvg'))
+        return _helpers._int_or_none(self._properties.get('readMsAvg'))
 
     @property
     def read_ms_max(self):
         """Union[int, None]: Milliseconds the slowest worker spent reading
         input.
         """
-        return _int_or_none(self._properties.get('readMsMax'))
+        return _helpers._int_or_none(self._properties.get('readMsMax'))
 
     @property
     def read_ratio_avg(self):
@@ -2702,14 +2701,14 @@ class QueryPlanEntry(object):
         """Union[int, None]: Milliseconds the average worker spent on CPU-bound
         processing.
         """
-        return _int_or_none(self._properties.get('computeMsAvg'))
+        return _helpers._int_or_none(self._properties.get('computeMsAvg'))
 
     @property
     def compute_ms_max(self):
         """Union[int, None]: Milliseconds the slowest worker spent on CPU-bound
         processing.
         """
-        return _int_or_none(self._properties.get('computeMsMax'))
+        return _helpers._int_or_none(self._properties.get('computeMsMax'))
 
     @property
     def compute_ratio_avg(self):
@@ -2732,14 +2731,14 @@ class QueryPlanEntry(object):
         """Union[int, None]: Milliseconds the average worker spent writing
         output data.
         """
-        return _int_or_none(self._properties.get('writeMsAvg'))
+        return _helpers._int_or_none(self._properties.get('writeMsAvg'))
 
     @property
     def write_ms_max(self):
         """Union[int, None]: Milliseconds the slowest worker spent writing
         output data.
         """
-        return _int_or_none(self._properties.get('writeMsMax'))
+        return _helpers._int_or_none(self._properties.get('writeMsMax'))
 
     @property
     def write_ratio_avg(self):
@@ -2760,12 +2759,12 @@ class QueryPlanEntry(object):
     @property
     def records_read(self):
         """Union[int, None]: Number of records read by this stage."""
-        return _int_or_none(self._properties.get('recordsRead'))
+        return _helpers._int_or_none(self._properties.get('recordsRead'))
 
     @property
     def records_written(self):
         """Union[int, None]: Number of records written by this stage."""
-        return _int_or_none(self._properties.get('recordsWritten'))
+        return _helpers._int_or_none(self._properties.get('recordsWritten'))
 
     @property
     def status(self):
@@ -2777,14 +2776,16 @@ class QueryPlanEntry(object):
         """Union[int, None]: Number of bytes written by this stage to
         intermediate shuffle.
         """
-        return _int_or_none(self._properties.get('shuffleOutputBytes'))
+        return _helpers._int_or_none(
+            self._properties.get('shuffleOutputBytes'))
 
     @property
     def shuffle_output_bytes_spilled(self):
         """Union[int, None]: Number of bytes written by this stage to
         intermediate shuffle and spilled to disk.
         """
-        return _int_or_none(self._properties.get('shuffleOutputBytesSpilled'))
+        return _helpers._int_or_none(
+            self._properties.get('shuffleOutputBytesSpilled'))
 
     @property
     def steps(self):
@@ -2828,31 +2829,31 @@ class TimelineEntry(object):
     def elapsed_ms(self):
         """Union[int, None]: Milliseconds elapsed since start of query
         execution."""
-        return _int_or_none(self._properties.get('elapsedMs'))
+        return _helpers._int_or_none(self._properties.get('elapsedMs'))
 
     @property
     def active_units(self):
         """Union[int, None]: Current number of input units being processed
         by workers, reported as largest value since the last sample."""
-        return _int_or_none(self._properties.get('activeUnits'))
+        return _helpers._int_or_none(self._properties.get('activeUnits'))
 
     @property
     def pending_units(self):
         """Union[int, None]: Current number of input units remaining for
         query stages active at this sample time."""
-        return _int_or_none(self._properties.get('pendingUnits'))
+        return _helpers._int_or_none(self._properties.get('pendingUnits'))
 
     @property
     def completed_units(self):
         """Union[int, None]: Current number of input units completed by
         this query."""
-        return _int_or_none(self._properties.get('completedUnits'))
+        return _helpers._int_or_none(self._properties.get('completedUnits'))
 
     @property
     def slot_millis(self):
         """Union[int, None]: Cumulative slot-milliseconds consumed by
         this query."""
-        return _int_or_none(self._properties.get('totalSlotMs'))
+        return _helpers._int_or_none(self._properties.get('totalSlotMs'))
 
 
 class UnknownJob(_AsyncJob):
